requirements.txt
requests
numpy
websocket-client
flask
flask-cors
schedule
